requirements.txt
SQLAlchemy~=2.0.36
python-dotenv~=1.0.1
mysql-connector-python~=9.1.0
requests~=2.32.3
beautifulsoup4~=4.12.3
emoji~=2.14.0
aiogram~=3.15.0
APScheduler~=3.10.4
selenium~=4.26.1
pyTelegramBotAPI~=4.24.0
uvicorn~=0.32.1
fastapi~=0.115.5